+++ src/chart/mod.rs
@@ -14,10 +14,12 @@ detailed description for each struct.
 
 mod builder;
 mod context;
+mod dual_coord;
 mod mesh;
 mod series;
 
 pub use builder::ChartBuilder;
 pub use context::{ChartContext, SeriesAnno};
+pub use dual_coord::DualCoordChartContext;
 pub use mesh::MeshStyle;
 pub use series::{SeriesLabelPosition, SeriesLabelStyle};